Revert "EndPoints_Marcelo"
This reverts commit da11584d0abc8a5d3dcae9730a1060263e664a33.

+++ API/Controllers/UnidadesController.cs
@@ -32,10 +32,10 @@ namespace API.Controllers
         {
             try
             {
-                IEnumerable<long> unidades = null;
+                IEnumerable<int> unidades = null;
                 if (!string.IsNullOrEmpty(ids))
                 {
-                    unidades = ids.Split(',').Select(x => Convert.ToInt64(x));
+                    unidades = ids.Split(',').Select(x => Convert.ToInt32(x));
                 }
 
                 return await _unidadesQueryService.GetAllAsync(page, take, unidades);
